move resources/ts to src
