--- resources/ts/core/tables/DataTableRow.vue
@@ -1,4 +1,0 @@
-<template>
-  <tr class="mdc-data-table__row"><slot></slot></tr>
-</template>
-
