--- resources/ts/core/tables/store.ts
@@ -1,3 +1,0 @@
-import { ref } from "vue";
-
-export const thUpdateKey = ref(0)
