--- resources/ts/store/index.ts
@@ -1,20 +1,0 @@
-import { createPinia, PiniaPlugin } from 'pinia'
-import { markRaw } from 'vue'
-import { router } from '../router'
-import 'pinia'
-import { Router } from 'vue-router'
-
-export const store = createPinia()
-
-const routerPlugin: PiniaPlugin = ({ store }) => ({
-    router: markRaw(router)
-})
-
-store.use(routerPlugin)
-
-// We have to type our new 'router' property so we can use it elsewhere with TS recognizing it
-declare module 'pinia' {
-  export interface PiniaCustomProperties {
-    get router(): Router
-  }
-}
