+++ src/core/tables/DataTableRow.vue
@@ -1,0 +1,5 @@
+<template>
+    <tr class="mdc-data-table__row">
+        <slot></slot>
+    </tr>
+</template>
