+++ src/core/tables/_index.scss
