+++ src/core/tables/store.ts
@@ -1,0 +1,3 @@
+import { ref } from "vue";
+
+export const thUpdateKey = ref(0);
